Adds support to env vars #5
With this update the user will be allowed to set the Harpoon service in three different flavors or even mix them to attend to their needs.

Default:
  - Bind address for client connections: 0.0.0.0
  - TCP port for client connections: 10608
  - Socket to connect to Docker daemon: unix:///var/run/docker.sock

Via command line flags (via --help flag):
  -bind string
      Bind address for client connection (default "0.0.0.0")
  -port string
      TCP port for client connection (default "10608")
  -socket string
      Socket location (default "unix:///var/run/docker.sock")

Via environment variables:
  HARPOON_BIND
      equivalent to -bind flag
  HARPOON_PORT
      equivalent to -port flag
  HARPOON_SOCKET
      equivalent to -socket flag

For example, to change the TCP port to a custom value you can use the flag -port:
  $ harpoond --port=12345

or you can use the environment variable:
  $ HARPOON_PORT=12345 harpoond

It's going to be useful to setup a Harpoon daemon instance running inside a container.